requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.8
filelock==3.20.3
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
